requirements.txt
Jinja2-3.1.5 
MarkupSafe-3.0.2 
Werkzeug-3.1.3 
blinker-1.9.0 
click-8.1.8 
colorama-0.4.6
flask-3.1.0
itsdangerous-2.2.0
python-dotenv-1.0.1
certifi-2025.1.31
charset-normalizer-3.4.1
idna-3.10
requests-2.32.3
urllib3-2.3.0
Flask-SQLAlchemy-3.1.1 
greenlet-3.1.1 
sqlalchemy-2.0.39 
typing-extensions-4.12.2
Flask-Migrate-4.1.0 
Mako-1.3.9 
alembic-1.15.1